Make a testing package and move service tests to the service package
